fixed mock helper funcs & some funcs on main after review
